🤫 Added silent switch
