+++ GoAwayEdge/GoAwayEdge.csproj
@@ -12,12 +12,17 @@
     <AssemblyName>GoAwayEdge</AssemblyName>
     <Company>Exploitox</Company>
     <Authors>valnoxy</Authors>
-    <Version>1.3.0.96</Version>
-    <Copyright>Copyright (c) 2018 - 2023 Exploitox. All rights reserved.</Copyright>
+    <Version>1.3.0.109</Version>
+    <Copyright>Copyright (c) 2018 - 2024 Exploitox. All rights reserved.</Copyright>
     <PackageProjectUrl>https://github.com/valnoxy/GoAwayEdge</PackageProjectUrl>
     <RepositoryUrl>https://github.com/valnoxy/GoAwayEdge</RepositoryUrl>
     <StartupObject>GoAwayEdge.App</StartupObject>
     <ApplicationIcon>GoAwayEdge.ico</ApplicationIcon>
+  	<AssemblyVersion>$(Version)</AssemblyVersion> 
+	<FileVersion>$(Version)</FileVersion> 
+	<AssemblyInformationalVersion>$(Version)</AssemblyInformationalVersion> 
+	<IncludeSourceRevisionInInformationalVersion>false</IncludeSourceRevisionInInformationalVersion> 
+	<AssemblyTitle>GoAwayEdge - Redirect Edge calls to your favorite browser!</AssemblyTitle> 
   </PropertyGroup>
   <ItemGroup>
     <None Remove="Assets\SetupBannerIcon.png" />
